Create "reservations" table for tracking reservations
Instead of using a field in each of the users and scooters tables to determine
this, make a table to track each reservation. This tracks when each reservation
is started and ended. The same scooter can be reserved multiple times.

Currently attempts to end a reservation beyond the first will not succeed.